# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management System documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "List of all products",
                "operationId": "58c1a6ac0aaf000641c821b802252026",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "sku": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "type": "integer"
                                                    },
                                                    "stock": {
                                                        "type": "integer"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "id": 1,
                                            "sku": "sit-consequuntur-rerum",
                                            "name": "sit consequuntur rerum",
                                            "price": "490.66",
                                            "stock": 99,
                                            "description": "Fugiat eum et ut. Autem totam sunt asperiores minima.",
                                            "created_at": "2024-06-26T07:51:10.000000Z"
                                        },
                                        {
                                            "id": 2,
                                            "sku": "aut-omnis-vitae",
                                            "name": "aut omnis vitae",
                                            "price": "604.06",
                                            "stock": 78,
                                            "description": "Deserunt sed nobis est sint minus. Voluptas ut ut dolores laudantium et tempora.",
                                            "created_at": "2024-06-26T07:51:10.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create new product.",
                "operationId": "0ec0d87f9b641174df0f28100f3b760d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "sku": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "string"
                                            },
                                            "stock": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "example": {
                                "sku": "lorem-ipsum",
                                "name": "Lorem Ipsum",
                                "description": "Voluptatum sequi odio sint dolorem consectetur nihil quasi.",
                                "price": "10.20",
                                "stock": 10
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "sku": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "type": "integer"
                                                },
                                                "stock": {
                                                    "type": "integer"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
        "/api/v1/products/{product}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get single product",
                "operationId": "d48579323ca79a12e52e810c6aa6c9ea",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "sku": {
                                                    "type": "string",
                                                    "example": "lorem-ipsum"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "name"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": "10.20"
                                                },
                                                "stock": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Voluptatum sequi odio sint dolorem consectetur nihil quasi."
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-06-26T07:51:10.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update product",
                "operationId": "aa160685e8ec32327a5192bf22e46ee8",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "sku": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "string"
                                            },
                                            "stock": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "example": {
                                "sku": "lorem-ipsum-1111",
                                "name": "Lorem Ipsum",
                                "description": "Voluptatum sequi odio sint dolorem consectetur nihil quasi.",
                                "price": "10.20",
                                "stock": 10
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "sku": {
                                                    "type": "string",
                                                    "example": "lorem-ipsum"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "name"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": "10.20"
                                                },
                                                "stock": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Voluptatum sequi odio sint dolorem consectetur nihil quasi."
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-06-26T07:51:10.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete product",
                "operationId": "b1e0fae4b2bf362c0d44c30da91100d8",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}